ReferenceField: show linearProgress bar only if some data are fetched
The linearProgress bar was displayed whereas there was no data to fetch.
Thus the bar was displayed for ever. The only way to not have the
progress bar forever, was to set allowEmpty to true.
However, this was removing completely the waiting bar.
And this is not a good idea, because users will not see that data are loading.

So now, if allowEmpty is set, linearProgress is shown while the data
are loaded, and if there is no data to fetch the bar is hidden.

The behavior is not changed if allowEmpty is not set.