sidebar: add backgroudColor property
The sideBar layout is defined by a paper component from material ui.
This paper is used in adminOnRest list views and we just want
set backgroundColor for the sidebar only and not for the list to.
now we add backgroundColor property inherited to the drawer color
property. We do this in order to change the backgroundColor of the
sideBar and to have unifyed theme for lateral menu.