sidebar: code review
We set drawer.color property to the paper component from material ui.
But we don't check if this property is undefined. Now we add if
drawer.color is defined we set it to the paper backgroundColor, in other
case we don't set this property. we do this in order to try setting
undefined value.