menuItemLink: add default style property
In the app the menuItemlink is defined with menuItem from material ui,
With material ui we can defined some style property but we can't change
text color property directely. Now we add default color property to the
menuItem of the sidebar. We do this in order to have a custom cleany
sidebar.