UI Components 3) List & Bulkactions
MenuButton: A generic component which renders a given button with a Menu.
BulkActionButton: A generic component which renders a button with a badge to indicate the number of selected items.
BulkActions: The List action menu component for executing bulk actions. It allows multiple BulkAction items.
BulkAction: Base component to be used for BulkAction items.
BulkActionCustom: A redux connected version of BulkAction, start point for creating custom BulkAction items.
BulkActionDelete: Basic composition of BulkActionCustom. This is the default action rendered in BulkActions.