Added new function - onBeforeAdd to SelectArrayInput. this was needed to be able to transfrom the new value to a different value based on some conditions and even to stop adding a new  value based on some conditions
Real usecase: choices are provided by server side, if user add new value with same letters but different letter case ("FrEe" against "free") it added a new value which we didnt want. to prevent that I added the onAddBefore so we could transform the "FrEe" to "free". also I was able to prevent duplication of same tag more than once.