[RFR] Use default colors for most buttons
We currently apply set the color prop of all buttons to `primary` which defeat the purpose of having a primary button.

This PR sets the default color of all buttons to `default`. It is still overridable by supplying the `color` prop.

Same logic is applied to the `DeleteButton` which should not stand out by default (it is currently always red). However, a new `danger` value is supported by its `color` prop in addition of the material-ui values for when it is needed (such as in a confirmation modal)