Export all from `TranslationContext` file
The `TranslationContext` instance is required so a component can be flagged
with the correct `static contextType = TranslationContext` value. Without
setting that (at least in React 16.7), the `this.context` value is fully
undefined and the component is not actually tied to the context.

By exporting it, consumer can simply set the context type and everything
works out of the box.