`sort` in <ReferenceArrayInput> doesn't work
Whatever I put in the `sort` props of the `<ReferenceArrayInput>` component is not taken into account, and it still filters the content by id.

The only difference I noticed is that by default, it sorts the referenced content by `id` DESC, and when I use the sort option it sorts by `id` ASC.

Reading the source code, it seems that the `useReferenceArrayInputController` doesn't do its job properly, but from a quick glance I didn't find the problem.

AFAIK, the bug exists in both the latest 3.x and 4.x versions.