Fix react-admin requires custom routers to be data routers
The reimplentation of the `warnWhenUnsavedChanges` feature with react-router's `useBlocker` created a breaking change. As `useBlocker` throws an error when not called in a DataRouter, and since we cannot call hooks conditionally, this meant that react-admin MUST be used inside a data router.

By moving the `warnWhenUnsavedChanges` logic from the `useAugmentedForm` hook to a custom component, we can now call the component conditionally - and avoid calling it when not inside a data router.