Allow overriding the error props in Input components
The basic idea is to be able to override when a helper text becomes an error. In our case we want to override the behaviour of react-admin to only consider a field as errored when the field is touched or submitted. We would _always_ like to show the error message.
Unfortunately, react-hook-form does not provide a way to set the internal touched state. However, when being able to override the error prop we can customize the behaviour ourselves.

The CheckboxGroupInput.tsx is also changed here, as the initial omitting of the error prop was introduced to fix a bug in this component. That's also why I moved the omitting of the prop to the local omit within the component.